// src/app/open-api/generated/model/tracking.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FileTrackingInfo } from './file-tracking-info';


export interface Tracking { 
    id: number;
    name: string;
    jobExecutionId?: string;
    status: Tracking.StatusEnum;
    flowType: Tracking.FlowTypeEnum;
    type: string;
    files: Array<FileTrackingInfo>;
    ok: Array<FileTrackingInfo>;
    ko: Array<FileTrackingInfo>;
    /**
     * Timestamp when the tracking entry was created
     */
    createdAt?: string;
    /**
     * Timestamp when the tracking entry was last updated
     */
    updatedAt?: string;
}
export namespace Tracking {
    export const StatusEnum = {
        Pending: 'PENDING',
        InProgress: 'IN_PROGRESS',
        Completed: 'COMPLETED',
        Cancelled: 'CANCELLED',
        Failed: 'FAILED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const FlowTypeEnum = {
        Import: 'import',
        Export: 'export'
    } as const;
    export type FlowTypeEnum = typeof FlowTypeEnum[keyof typeof FlowTypeEnum];
}